kata-runtime: kata-check changes to Enable Darwin build
kata-check is very much Linux specific, on both ARM and x86.
We split the OS agnostic part from the Linux specific bits and
keep the former in kata-check.go while the latter go to
kata-check_linux.go.

And we define Darwin and architecture specific skeleton files
for kata-check on Darwin (no real checks added yet for Darwin).
